pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "articlewriter"
version = "1.0.0"
description = "Automated journal-ready research article generation in APA 7"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "python-docx>=1.1.0",
    "reportlab>=4.0.0",
    "openai>=1.6.0",
    "anthropic>=0.18.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "streamlit>=1.29.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "bibtexparser>=1.4.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-asyncio>=0.23.0"]
bertopic = ["bertopic>=0.16.0", "sentence-transformers>=2.2.0"]
api = ["fastapi>=0.109.0", "uvicorn[standard]>=0.27.0"]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
articlewriter = "articlewriter.cli:app"
